move v1 profiler callback to callbacks package
