+++ src/llamafactory/v1/core/base_trainer.py
@@ -39,6 +39,7 @@ from ..utils import logging
 from ..utils.callbacks import (
     CallbackHandler,
     LoggingCallback,
+    ProfilerCallback,
     TrainerCallback,
     TrainerState,
 )
@@ -126,8 +127,6 @@ class BaseTrainer:
         # Callbacks
         self.callback_handler = CallbackHandler([LoggingCallback()], trainer=self)
         if getattr(self.args, "enable_torch_profiler", False):
-            from ..accelerator.profiler import ProfilerCallback
-
             self.callback_handler.add_callback(ProfilerCallback(self.args))
 
         for cb in callbacks or []:
